implement suggested changes
