+++ src/components/PeopleTable/PeopleTable.tsx
@@ -1,42 +1,84 @@
+import { useEffect, useState } from 'react';
 import { Person } from '../../types';
+import { Loader } from '../Loader';
 import { PersonRow } from '../PersonRow';
+import { getPeople } from '../../api';
 
-type Props = {
-  people?: Person[] | null;
-};
+const STATUS = {
+  resolved: 'resolved',
+  rejected: 'rejected',
+  idle: 'idle',
+  pending: 'pending',
+} as const;
 
-export const PeopleTable = ({ people }: Props) => {
+type Status = (typeof STATUS)[keyof typeof STATUS];
+
+export const PeopleTable = () => {
+  const [status, setStatus] = useState<Status>('idle');
+  const [people, setPeople] = useState<Person[] | null>(null);
+
+  const loadPeople = () => {
+    return getPeople()
+      .then(data => {
+        setPeople(data);
+        setStatus('resolved');
+      })
+      .catch(() => setStatus('rejected'));
+  };
+
+  useEffect(() => {
+    setStatus('pending');
+    loadPeople();
+  }, []);
+
   return (
     <>
       <h1 className="title">People Page</h1>
-      <table
-        data-cy="peopleTable"
-        className="
+      {people?.length !== 0 && status === 'resolved' && (
+        <table
+          data-cy="peopleTable"
+          className="
                     table
                     is-striped
                     is-hoverable
                     is-narrow
                     is-fullwidth"
-      >
-        <thead>
-          <tr>
-            <th>Name</th>
-            <th>Sex</th>
-            <th>Born</th>
-            <th>Died</th>
-            <th>Mother</th>
-            <th>Father</th>
-          </tr>
-        </thead>
+        >
+          <thead>
+            <tr>
+              <th>Name</th>
+              <th>Sex</th>
+              <th>Born</th>
+              <th>Died</th>
+              <th>Mother</th>
+              <th>Father</th>
+            </tr>
+          </thead>
 
-        <tbody>
-          {people?.map(person => {
-            return (
-              <PersonRow key={person.name} people={people} person={person} />
-            );
-          })}
-        </tbody>
-      </table>
+          <tbody>
+            {people?.map(person => {
+              return (
+                <PersonRow key={person.name} people={people} person={person} />
+              );
+            })}
+          </tbody>
+        </table>
+      )}
+      <div className="block">
+        <div className="box table-container">
+          {status === 'pending' && <Loader />}
+
+          {status === 'rejected' && (
+            <p data-cy="peopleLoadingError" className="has-text-danger">
+              Something went wrong
+            </p>
+          )}
+
+          {people?.length === 0 && (
+            <p data-cy="noPeopleMessage">There are no people on the server</p>
+          )}
+        </div>
+      </div>
     </>
   );
 };
